gas dimension trace tests: split into separate files
it was becoming far too unweildy in one file for navigation
